projet_mooc/requirements.txt
absl-py==2.1.0
alembic==1.13.1
altair==5.3.0
aniso8601==9.0.1
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.1
blinker==1.8.2
blis==0.7.11
boto3==1.34.118
botocore==1.34.118
c2xg==2.2
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
corpus-similarity==1.1
credentials==1.1
cycler==0.12.1
cymem==2.0.8
cytoolz==0.12.3
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.6.1
docker==7.1.0
emoji==1.7.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.52.4
ftfy==6.2.0
gast==0.5.4
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.64.1
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
importlib-metadata==7.0.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
kmedoids==0.5.1
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
Mako==1.3.5
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mlflow==2.13.0
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.45
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.7.2
pyparsing==3.1.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
querystring-parser==1.2.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.2
stop-words==2018.7.23
streamlit==1.35.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
textanalytics==1.1
thinc==8.2.3
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.9.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.3
watchdog==4.0.1
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.3
wordcloud==1.9.3
wrapt==1.16.0
xgboost==2.0.3
zipp==3.19.0
